+++ playgrounds/qwik/src/routes/special/index.tsx
@@ -7,6 +7,7 @@ import {
   FileInput,
   FormFooter,
   FormHeader,
+  RadioGroup,
   Select,
   Slider,
   TextInput,
@@ -19,6 +20,7 @@ const SpecialFormSchema = v.object({
     array: v.array(v.string()),
     boolean: v.optional(v.boolean(), false),
   }),
+  radio: v.optional(v.string()),
   select: v.object({
     array: v.array(v.string()),
     string: v.optional(v.string()),
@@ -86,7 +88,7 @@ export default component$(() => {
               render$={(field) => (
                 <Checkbox
                   {...field.props}
-                  class="!p-0"
+                  class="p-0!"
                   label={label}
                   value={value}
                   input={createComputed$(() =>
@@ -108,6 +110,24 @@ export default component$(() => {
               input={field.input}
               errors={field.errors}
               label="Checkbox boolean"
+            />
+          )}
+        />
+
+        <Field
+          of={specialForm}
+          path={['radio']}
+          render$={(field) => (
+            <RadioGroup
+              {...field.props}
+              label="Radio group"
+              options={[
+                { label: 'Option 1', value: 'option_1' },
+                { label: 'Option 2', value: 'option_2' },
+                { label: 'Option 3', value: 'option_3' },
+              ]}
+              input={field.input}
+              errors={field.errors}
             />
           )}
         />