+++ playgrounds/solid/src/components/Radio.tsx
@@ -1,0 +1,27 @@
+import type { FieldElementProps } from '@formisch/solid';
+import { splitProps } from 'solid-js';
+
+interface RadioProps extends FieldElementProps {
+  class?: string;
+  label: string;
+  value: string;
+  checked: boolean;
+}
+
+/**
+ * Simple radio button input. Should be used inside a RadioGroup component.
+ */
+export function Radio(props: RadioProps) {
+  const [, inputProps] = splitProps(props, ['class', 'label']);
+
+  return (
+    <label class="flex cursor-pointer items-center space-x-3 font-medium select-none md:text-lg lg:text-xl">
+      <input
+        {...inputProps}
+        class="h-4 w-4 cursor-pointer lg:h-5 lg:w-5"
+        type="radio"
+      />
+      <span>{props.label}</span>
+    </label>
+  );
+}