+++ playgrounds/solid/src/routes/special/index.tsx
@@ -6,6 +6,7 @@ import {
   FileInput,
   FormFooter,
   FormHeader,
+  RadioGroup,
   Select,
   Slider,
   TextInput,
@@ -19,6 +20,7 @@ const SpecialFormSchema = v.object({
     array: v.array(v.string()),
     boolean: v.optional(v.boolean(), false),
   }),
+  radio: v.optional(v.string()),
   select: v.object({
     array: v.array(v.string()),
     string: v.optional(v.string()),
@@ -84,7 +86,7 @@ export default function SpecialPage() {
                   {(field) => (
                     <Checkbox
                       {...field.props}
-                      class="!p-0"
+                      class="p-0!"
                       label={label}
                       value={value}
                       input={field.input.includes(value)}
@@ -103,6 +105,22 @@ export default function SpecialPage() {
                 input={field.input}
                 errors={field.errors}
                 label="Checkbox boolean"
+              />
+            )}
+          </Field>
+
+          <Field of={specialForm} path={['radio']}>
+            {(field) => (
+              <RadioGroup
+                {...field.props}
+                label="Radio group"
+                options={[
+                  { label: 'Option 1', value: 'option_1' },
+                  { label: 'Option 2', value: 'option_2' },
+                  { label: 'Option 3', value: 'option_3' },
+                ]}
+                input={field.input}
+                errors={field.errors}
               />
             )}
           </Field>