+++ packages/core/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Fix radio button value handling in `getElementInput` (pull request #39)
+
 ## v0.4.2 (November 25, 2025)
 
 - Fix `isDirty` to handle `null` like `undefined` for empty string and `NaN` comparisons (pull request #40)